% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{autoreg}
\alias{autoreg}
\title{Generate Time-Varying Autoregressive covariate}
\usage{
autoreg(n, rho, sigma)
}
\arguments{
\item{n}{Integer. Number of values to generate after discarding.}

\item{rho}{Numeric. Autoregressive coefficient.}

\item{sigma}{Numeric. Standard deviation of the noise term.}
}
\value{
A numeric vector containing the generated time series.
}
\description{
Generates a time-varying autoregressive series and discards the first half of the generated series.
}
\examples{
ts <- autoreg(100, 0.8, 0.1)
plot(ts, type = "l")

}
