%
%  Copyright (C) 2005 Friedrich Leisch
%  $Id: birth.Rd 252 2018-09-17 08:40:24Z gruen $
%
\name{birth}
\alias{birth}
\docType{data}
\title{Birth and Death Rates}
\description{
  Birth and death rates for 70 countries.
}
\usage{data(birth)}
\format{
  A data frame with 70 observations on the following 2 variables.
  \describe{
    \item{\code{birth}}{Birth rate (in percent).}
    \item{\code{death}}{Death rate (in percent).}
  }
}
%\source{
%  \url{http://www.uni-koeln.de/themen/statistik}
%}
\references{
 John A. Hartigan: Clustering Algorithms. Wiley, New York, 1975.
}
\keyword{datasets}
