%
%  Copyright (C) 2005 Friedrich Leisch
%  $Id: dentitio.Rd 3 2013-06-12 10:06:43Z leisch $
%
\name{dentitio}
\alias{dentitio}
\docType{data}
\title{Dentition of Mammals}
\description{
  Mammal's teeth divided into the 4 groups: incisors, canines, premolars
  and molars. 
}
\usage{data(dentitio)}
\format{
  A data frame with 66 observations on the following 8 variables.
  \describe{
    \item{top.inc:}{top incisors}
    \item{bot.inc:}{bottom incisors}
    \item{top.can:}{top canines}
    \item{bot.can:}{bottom canines}
    \item{top.pre:}{top premolars}
    \item{bot.pre:}{bottom premolars}
    \item{top.mol:}{top molars}
    \item{bot.mol:}{bottom molars}
  }
}
\source{
  \url{http://www.uni-koeln.de/themen/statistik}
}
\references{
 John A. Hartigan: Clustering Algorithms. Wiley, New York, 1975.
}
\keyword{datasets}
