% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsDist.R
\name{dist.parallel}
\alias{dist.parallel}
\title{Calculate the Distance Matrix of Multiple Functional or Time-Series Curves}
\usage{
dist.parallel(tsList, distFunc = tsdiss.euclidean, cores = NUM_CORES)
}
\arguments{
\item{tsList}{A list of multiple time-series curves.}

\item{distFunc}{A distance function, such as \code{tsdiss.euclidean}, \code{tsdiss.manhattan}, or \code{tsdiss.correlation}.}

\item{cores}{The number of CPU cores to use for parallel computing. Defaults to the maximum available cores minus one.}
}
\description{
Computes pairwise distances between multiple time-series curves using a specified distance function, with optional parallel computing support.
}
\examples{
tsList <- list(rnorm(100), rnorm(100))
distMat <- dist.parallel(tsList, distFunc = tsdiss.euclidean, cores = 2)
print(distMat)

}
\keyword{distance}
\keyword{matrix}
