% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsDist.R
\name{dist.parallel2.compute}
\alias{dist.parallel2.compute}
\title{Compute Pairwise Distance Between Two Time-Series Curves}
\usage{
dist.parallel2.compute(x, y, tsList, distFunc)
}
\arguments{
\item{x}{Index of the first time-series curve in the list.}

\item{y}{Index of the second time-series curve in the list.}

\item{tsList}{A list of time-series curves.}

\item{distFunc}{A compiled distance function.}
}
\value{
A numeric distance value between two time-series curves.
}
\description{
Computes the distance between two time-series curves using a specified distance function.
}
