% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_build.R
\name{dist_percentile}
\alias{dist_percentile}
\title{Compute Distance Percentile Threshold}
\usage{
dist_percentile(distance_matrix, percentile_value)
}
\arguments{
\item{distance_matrix}{A square distance matrix.}

\item{percentile_value}{A numeric value between 0 and 1 representing the desired percentile threshold.}
}
\value{
A numeric value representing the distance threshold at the given percentile.
}
\description{
Computes a distance threshold based on a given percentile from a square distance matrix.
}
\details{
\itemize{
\item The function extracts the upper triangle of the distance matrix (excluding the diagonal).
\item Computes the percentile threshold based on the specified percentile value.
}
}
\examples{
set.seed(123)
distance_matrix <- matrix(runif(100), nrow = 10)
threshold <- dist_percentile(distance_matrix, percentile_value = 0.5)
print(threshold)

}
