% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmi.R
\name{Hc}
\alias{Hc}
\title{Entropy of a Partition}
\usage{
Hc(C, n)
}
\arguments{
\item{C}{A list representing a partition.}

\item{n}{An integer representing the set size.}
}
\value{
A numeric value representing the entropy of the partition.
}
\description{
Computes the entropy of a partition, representing the uncertainty within the partition.
}
