% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmi.R
\name{Icd}
\alias{Icd}
\title{Mutual Information Between Two Partitions}
\usage{
Icd(C, D, n)
}
\arguments{
\item{C}{A list representing one partition.}

\item{D}{A list representing another partition.}

\item{n}{An integer representing the set size.}
}
\value{
A numeric value representing the mutual information between partitions.
}
\description{
Computes the mutual information shared between two partitions.
}
