% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsknot.R
\name{lsest2}
\alias{lsest2}
\title{Varying Coefficient Estimation}
\usage{
lsest2(inialp, Sam, y, kC, kA, ord, delta)
}
\arguments{
\item{inialp}{Initial estimates from the first step.}

\item{Sam}{A data frame of predictors.}

\item{y}{A vector of responses.}

\item{kC}{Number of knots for the time component.}

\item{kA}{Number of knots for covariate components.}

\item{ord}{Integer. Order of the B-spline basis.}

\item{delta}{Regularization parameter for ridge regression.}
}
\value{
A list containing estimated functions and model metrics:
\item{estfun}{Matrix of estimated functions.}
\item{rss2}{Residual sum of squares.}
\item{paranum}{Number of parameters in the model.}
}
\description{
Performs the second-step estimation of varying-coefficient
functions using spline regression.
}
\examples{
Sam <- data.frame(matrix(rnorm(300), ncol = 3))
y <- rnorm(100)
inialp <- matrix(rnorm(300), ncol = 3)
res <- lsest2(inialp, Sam, y, kC = 5, kA = 3, ord = 3, delta = 0.01)
str(res)

}
