% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsknot.R
\name{lsestIni}
\alias{lsestIni}
\title{Initial Estimation of Varying Coefficient Model}
\usage{
lsestIni(Sam, y, x01, x02, indt, hC, hA, ord, delta)
}
\arguments{
\item{Sam}{A data frame of predictors.}

\item{y}{A vector of responses.}

\item{x01}{Prediction points for the first covariate.}

\item{x02}{Prediction points for the second covariate.}

\item{indt}{Index of a reference time point for normalization.}

\item{hC}{Number of knots for the time component.}

\item{hA}{Number of knots for covariate components.}

\item{ord}{Integer. Order of the B-spline basis.}

\item{delta}{Regularization parameter for ridge regression.}
}
\value{
A list containing initial coefficient estimates and model metrics:
\item{inialp}{Matrix of initial estimates.}
\item{rss2}{Residual sum of squares.}
\item{paranum}{Number of parameters in the model.}
}
\description{
Computes initial estimates for varying-coefficient functions
using a spline-based regression approach.
}
\examples{
Sam <- data.frame(matrix(rnorm(300), ncol = 3))
y <- rnorm(100)
res <- lsestIni(Sam, y, x01 = rnorm(100), x02 = rnorm(100), indt = 1, hC = 5, hA = 3, ord = 3, delta = 0.01)
str(res)

}
