% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsknot.R
\name{lsknot}
\alias{lsknot}
\title{Select Optimal Knots for Varying Coefficient Additive Model}
\usage{
lsknot(Sam, y, x01, x02, indt, ord, delta)
}
\arguments{
\item{Sam}{A data frame of predictors.}

\item{y}{A vector of responses.}

\item{x01}{A vector of prediction points for the first covariate.}

\item{x02}{A vector of prediction points for the second covariate.}

\item{indt}{Index of a reference time point for normalizing estimates.}

\item{ord}{Integer. Order of the B-spline basis.}

\item{delta}{Regularization parameter for ridge regression.}
}
\value{
A vector of selected knot values for the model.
}
\description{
Selects optimal knots for the initial and varying-coefficient
estimation steps in a spline-based regression model.
}
\examples{

}
