% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_build.R
\name{network_build}
\alias{network_build}
\title{Build a Network from Distance Matrix}
\usage{
network_build(distance_matrix, percentile_value = 0.3, normalize = TRUE)
}
\arguments{
\item{distance_matrix}{A square distance matrix (n x n), where n represents the number of nodes.}

\item{percentile_value}{A numeric value between 0 and 1 representing the percentile threshold for edge formation.}

\item{normalize}{A logical value indicating whether to normalize the distance matrix before processing. Default is \code{TRUE}.}
}
\value{
An undirected graph object of class \code{igraph}.
}
\description{
Constructs an undirected network graph based on a distance matrix, using a specified distance percentile threshold.
}
\details{
\itemize{
\item The function normalizes the distance matrix (if \code{normalize = TRUE}) and identifies edges based on a percentile threshold.
\item Nodes with distances below the given threshold are connected by edges in the resulting graph.
}
}
\examples{
set.seed(123)
distance_matrix <- matrix(runif(100), nrow = 10)
network <- network_build(distance_matrix, percentile_value = 0.5)
plot(network)

}
