% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_build.R
\name{network_clust}
\alias{network_clust}
\title{Perform Network Clustering}
\usage{
network_clust(net)
}
\arguments{
\item{net}{An igraph object representing the network to be clustered.}
}
\value{
An object of class \code{communities}, representing the detected community structure.
}
\description{
Clusters a network into communities.
}
