% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmi.R
\name{NMI}
\alias{NMI}
\title{Normalized Mutual Information (NMI) Between Two Partitions}
\usage{
NMI(C, D, n)
}
\arguments{
\item{C}{A list representing one partition of a set S. Each element in the list is a subset of indices from the set.}

\item{D}{A list representing another partition of the same set S.}

\item{n}{An integer representing the cardinality (size) of the set S.}
}
\value{
A numeric value representing the NMI score between the two partitions.
}
\description{
Computes the Normalized Mutual Information (NMI) to measure the similarity between two partitions of a set.
}
\details{
\itemize{
\item \strong{Normalized Mutual Information (NMI):} Measures the shared information between two partitions normalized by their entropies.
\item NMI ranges from 0 to 1, where 1 indicates identical partitions and 0 indicates completely independent partitions.
\item The calculation is based on mutual information (\code{Icd}) and entropy (\code{Hc}) of each partition.
}
}
\examples{
n <- 20
C <- list(1:(n/2+1), (n/2+2):n, (n+1):(2*n))
D <- list(1:(n/2), (n/2+1):n, (n+1):(2*n))
res <- NMI(C, D, 2 * n)
print(res)
}
\keyword{clustering}
\keyword{evaluation,}
\keyword{information,}
\keyword{mutual}
\keyword{normalized}
\keyword{partition}
\keyword{similarity}
