%
%  Copyright (C) 2005 Friedrich Leisch
%  $Id: nutrient.Rd 252 2018-09-17 08:40:24Z gruen $
%
\name{nutrient}
\alias{nutrient}
\docType{data}
\title{Nutrients in Meat, Fish and Fowl}
\description{
  The data set contains the measurements of nutrients in several types
  of meat, fish and fowl.
}
\usage{data(nutrient)}
\format{
  A data frame with 27 observations on the following 5 variables.
  \describe{
    \item{\code{energy}}{Food energy (calories).}
    \item{\code{protein}}{Protein (grams).}
    \item{\code{fat}}{Fat (grams).}
    \item{\code{calcium}}{calcium (milli grams).}
    \item{\code{iron}}{Iron (milli grams).}
  }
}
%\source{
%  \url{http://www.uni-koeln.de/themen/statistik}
%}
\references{
 John A. Hartigan: Clustering Algorithms. Wiley, New York, 1975.
}
\keyword{datasets}
