%
%  Copyright (C) 2005 Friedrich Leisch
%  $Id: parameters.Rd 2479 2006-03-01 15:04:04Z leisch $
%
\name{parameters}
\alias{parameters}
\alias{parameters,kccasimple-method}
\title{Get Centroids from KCCA Object}
\description{
  Returns the matrix of centroids of a fitted object of class \code{"kcca"}.
}
\usage{
\S4method{parameters}{kccasimple}(object, ...)
}
\arguments{
  \item{object}{fitted object.}
  \item{\dots}{currently not used.}
}
\author{Friedrich Leisch}
\keyword{cluster}
