% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate Data for Dynamic Varying Coefficient Additive Model (VCAM)}
\usage{
simulate_data(n, T, x_mat)
}
\arguments{
\item{n}{Integer. Number of units.}

\item{T}{Integer. Number of time points.}

\item{x_mat}{Matrix. Covariate values for simulation.}
}
\value{
A list containing simulated data components:
\item{ymat}{Matrix of response variable values.}
\item{ylag}{Matrix of lagged response variable values.}
\item{umat}{Matrix of U values.}
\item{xmat}{Matrix of covariate values.}
\item{alpha0}{Matrix of intercept terms.}
\item{alp1}{Matrix of group-specific varying coefficients (part 1).}
\item{alp2}{Matrix of group-specific varying coefficients (part 2).}
\item{beta1}{Matrix of additive function beta1 values.}
\item{beta2}{Matrix of additive function beta2 values.}
}
\description{
Simulates data for a dynamic varying coefficient additive model.
}
\examples{
x_mat <- matrix(rnorm(100), nrow = 10, ncol = 10)
result <- simulate_data(10, 5, x_mat)
str(result)

}
