% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsDist.R
\name{tsdiss.correlation}
\alias{tsdiss.correlation}
\title{Correlation-Based Distance Between Time-Series Curves}
\usage{
tsdiss.correlation(ts1, ts2)
}
\arguments{
\item{ts1}{First time-series curve.}

\item{ts2}{Second time-series curve.}
}
\value{
A numeric distance value based on correlation.
}
\description{
Computes the distance between two time-series curves based on their correlation.
}
