% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsDist.R
\name{tsdiss.euclidean}
\alias{tsdiss.euclidean}
\title{Euclidean Distance Between Time-Series Curves}
\usage{
tsdiss.euclidean(ts1, ts2)
}
\arguments{
\item{ts1}{First time-series curve.}

\item{ts2}{Second time-series curve.}
}
\value{
A numeric Euclidean distance value.
}
\description{
Computes the Euclidean distance between two time-series curves.
}
