% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsDist.R
\name{tsdiss.infiniteNorm}
\alias{tsdiss.infiniteNorm}
\title{Infinite Norm Distance Between Time-Series Curves}
\usage{
tsdiss.infiniteNorm(ts1, ts2)
}
\arguments{
\item{ts1}{First time-series curve.}

\item{ts2}{Second time-series curve.}
}
\value{
A numeric infinite norm distance value.
}
\description{
Computes the infinite norm (maximum absolute difference) distance between two time-series curves.
}
