% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsDist.R
\name{tsdiss.manhattan}
\alias{tsdiss.manhattan}
\title{Manhattan Distance Between Time-Series Curves}
\usage{
tsdiss.manhattan(ts1, ts2)
}
\arguments{
\item{ts1}{First time-series curve.}

\item{ts2}{Second time-series curve.}
}
\value{
A numeric Manhattan distance value.
}
\description{
Computes the Manhattan (L1 norm) distance between two time-series curves.
}
