\name{achieve}
\alias{achieve}
\docType{data}
\title{Achievement Test Scores for New Haven Schools}
\description{
  Measurements at the beginning of the 4th grade (when the national
  average is 4.0) and of the 6th grade in 25 schools in New Haven. 
}
\usage{data(achieve)}
\format{
  A data frame with 25 observations on the following 4 variables.
  \describe{
    \item{read4:}{4th grade reading}
    \item{arith4:}{4th grade arithmetic}
    \item{read6:}{6th grade reading}
    \item{arith6:}{6th grade arithmetic}
  }
}
\source{
  \url{http://www.uni-koeln.de/themen/statistik}
}
\references{
 John A. Hartigan: Clustering Algorithms. Wiley, New York, 1975.
}
\keyword{datasets}
