\name{birth}
\alias{birth}
\docType{data}
\title{Birth and Death Rates}
\description{
  Birth and death rates for 70 countries.
}
\usage{data(birth)}
\format{
  A data frame with 70 observations on the following 2 variables.
  \describe{
    \item{birth:}{birth rate (in percent)}
    \item{death:}{death rate (in percent)}
  }
}
\source{
  \url{http://www.uni-koeln.de/themen/statistik}
}
\references{
 John A. Hartigan: Clustering Algorithms. Wiley, New York, 1975.
}
\keyword{datasets}
