%
%  Copyright (C) 2005 Friedrich Leisch
%  $Id: dentitio.Rd 252 2018-09-17 08:40:24Z gruen $
%
\name{dentitio}
\alias{dentitio}
\docType{data}
\title{Dentition of Mammals}
\description{
  Mammal's teeth divided into the 4 groups: incisors, canines, premolars
  and molars. 
}
\usage{data(dentitio)}
\format{
  A data frame with 66 observations on the following 8 variables.
  \describe{
    \item{\code{top.inc}}{Top incisors.}
    \item{\code{bot.inc}}{Bottom incisors.}
    \item{\code{top.can}}{Top canines.}
    \item{\code{bot.can}}{Bottom canines.}
    \item{\code{top.pre}}{Top premolars.}
    \item{\code{bot.pre}}{Bottom premolars.}
    \item{\code{top.mol}}{Top molars.}
    \item{\code{bot.mol}}{Bottom molars.}
  }
}
%\source{
%  \url{http://www.uni-koeln.de/themen/statistik}
%}
\references{
 John A. Hartigan: Clustering Algorithms. Wiley, New York, 1975.
}
\keyword{datasets}
