%
%  Copyright (C) 2005 Friedrich Leisch
%  $Id: milk.Rd 3 2013-06-12 10:06:43Z leisch $
%
\name{milk}
\alias{milk}
\docType{data}
\title{Milk of Mammals}
\description{
  The data set contains the ingredients of mammal's milk of 25 animals.
}
\usage{data(milk)}
\format{
  A data frame with 25 observations on the following 5 variables (all in
  percent). 
  \describe{
    \item{water:}{water}
    \item{protein:}{protein}
    \item{fat:}{fat}
    \item{lactose:}{lactose}
    \item{ash:}{ash}
  }
}
%\source{
%  \url{http://www.uni-koeln.de/themen/statistik}
%}
\references{
 John A. Hartigan: Clustering Algorithms. Wiley, New York, 1975.
}
\keyword{datasets}
