%
%  Copyright (C) 2005 Friedrich Leisch
%  $Id: milk.Rd 252 2018-09-17 08:40:24Z gruen $
%
\name{milk}
\alias{milk}
\docType{data}
\title{Milk of Mammals}
\description{
  The data set contains the ingredients of mammal's milk of 25 animals.
}
\usage{data(milk)}
\format{
  A data frame with 25 observations on the following 5 variables (all in
  percent). 
  \describe{
    \item{\code{water}}{Water.}
    \item{\code{protein}}{Protein.}
    \item{\code{fat}}{Fat.}
    \item{\code{lactose}}{Lactose.}
    \item{\code{ash}}{Ash.}
  }
}
%\source{
%  \url{http://www.uni-koeln.de/themen/statistik}
%}
\references{
 John A. Hartigan: Clustering Algorithms. Wiley, New York, 1975.
}
\keyword{datasets}
