\name{nutrient}
\alias{nutrient}
\docType{data}
\title{Nutrients in Meat, Fish and Fowl}
\description{
  The data set contains the measurements of nutrients in several types
  of meat, fish and fowl.
}
\usage{data(nutrient)}
\format{
  A data frame with 27 observations on the following 5 variables.
  \describe{
    \item{energy:}{food energy (calories)}
    \item{protein:}{protein (grams)}
    \item{fat:}{fat (grams)}
    \item{calcium:}{calcium (milli grams)}
    \item{iron:}{iron (milli grams)}
  }
}
\source{
  \url{http://www.uni-koeln.de/themen/statistik}
}
\references{
 John A. Hartigan: Clustering Algorithms. Wiley, New York, 1975.
}
\keyword{datasets}
