%
%  Copyright (C) 2005 Friedrich Leisch
%  $Id: parameters.Rd 252 2018-09-17 08:40:24Z gruen $
%
\name{parameters}
\alias{parameters}
\alias{parameters,kccasimple-method}
\title{Get Centroids from KCCA Object}
\description{
  Returns the matrix of centroids of a fitted object of class \code{"kcca"}.
}
\usage{
\S4method{parameters}{kccasimple}(object, ...)
}
\arguments{
  \item{object}{Fitted object.}
  \item{\dots}{Currently not used.}
}
\author{Friedrich Leisch}
\keyword{cluster}
