%
%  Copyright (C) 2005 Friedrich Leisch
%  $Id: parameters.Rd 1849 2005-10-10 06:15:57Z leisch $
%
\name{parameters}
\alias{parameters}
\alias{parameters,kcca-method}
\title{Get Centroids from KCCA Object}
\description{
  Returns the matrix of centroids of a fitted object of class \code{"kcca"}.
}
\usage{
\S4method{parameters}{kcca}(object, ...)
}
\arguments{
  \item{object}{fitted object.}
  \item{\dots}{currently not used.}
}
\author{Friedrich Leisch}
\keyword{cluster}
